% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_category.R
\name{rename_category}
\alias{rename_category}
\title{Rename a category of \code{stylo} data for \code{stylo2gg}.}
\usage{
rename_category(df, before, after)
}
\arguments{
\item{df}{An object saved from running \strong{\code{stylo}}
on a corpus}

\item{before}{A string representing the name of the category 
in the \code{stylo} object}

\item{after}{A string representing the new name to be used 
for this category}
}
\description{
Rename a category of \code{stylo} data for \code{stylo2gg}.
}
\details{
This change will not overwrite values in the originating 
\code{df} object.
}
\examples{
\dontrun{
my_data <- stylo()
my_data \%>\% rename_category("NA", "unknown")
}

}
