% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylo2gg.R
\name{stylo2gg}
\alias{stylo2gg}
\title{Visualize \code{stylo} data with \code{ggplot2}.}
\usage{
stylo2gg(
  df,
  viz,
  features,
  num.features,
  top.loadings,
  select.loadings,
  pc.x = 1,
  pc.y = 2,
  title = NULL,
  caption = FALSE,
  count.labels = FALSE,
  legend,
  black = NULL,
  highlight = NULL,
  labeling,
  classing,
  shapes = FALSE,
  invert.x = FALSE,
  invert.y = FALSE,
  scaling,
  distance.measure,
  linkage,
  horiz = TRUE,
  axis.labels = FALSE,
  highlight.nudge,
  highlight.single,
  show.zero,
  highlight.box = NULL,
  withholding,
  loadings.spacer = "_",
  loadings.line.color = "lightgray",
  loadings.word.color = "darkgray",
  loadings.upper = FALSE,
  plaintext = TRUE
)
}
\arguments{
\item{df}{An object saved from running \strong{\code{stylo}}
on a corpus}

\item{viz}{A choice of visualizations, either \code{"pca"}
for principal components analysis or \code{"hc"} for
hierarchical clustering; alternatively, using
\code{"PCR"}, \code{"PCV"}, or \code{"CA"}---all
inherited from \code{stylo}---will reset a number of
defaults.}

\item{features}{A vector containing a selection of features
to consider for analysis. This option is useful for
replicating a previous analysis.}

\item{num.features}{The number of features to be used for an
analysis. By default, \code{stylo}'s settings are used, but
it is easy here limit the number to a smaller set, ordered
by frequency}

\item{top.loadings}{The number of features to show as
vectors in a principal components analysis. By default,
loadings are not shown unless \code{stylo}'s setting for
\code{pca.visual.flavour} is set to \code{"loadings"}; at
this time, it defaults to the full number of features. It's
probably most revealing to choose a smaller number, so that
only the most significant features are plotted.}

\item{select.loadings}{A list element, with items indicating
either the nearest location of a selected feature or the names
of these features. The location can be shown in three ways: 1.
with coordinates in the PCA space, e.g. \code{c(1,2)}; 2. as
the number of a category from which to derive an average
location, e.g. \code{4}; 3. as the name of a category or some
other element of the original text's filename from which to
derive an average location, e.g. \code{"hamilton"}. The name of
a feature is the fourth option: 4. using a call with the
\code{word} function; for example, to show the word
"undershirt," the list item would be
\code{call("word", "undershirt")}. Multiple types of items can
be combined in one list:
\code{select.loadings = list(c(1,2), 4, "hamilton", call("word", "undershirt"))}.}

\item{pc.x}{Identifies the principal component to be placed 
on the X-axis. Defaults to \code{1}.}

\item{pc.y}{Identifies the principal component to be placed 
on the Y-axis. Defaults to \code{2}.}

\item{title}{The title that will go on the top of a chart.
This value is inherited from \code{stylo} where possible. To
remove a title, set it to \code{NULL} or to an empty
set \strong{""}.}

\item{caption}{Defaults to \code{FALSE}, except with
certain \strong{\code{viz}} settings. Change to toggle
metadata at the bottom of a visualization.}

\item{count.labels}{Toggle (TRUE / FALSE) to show or hide counting numbers at the beginning of labels on a dendrogram. Useful for manually setting a \strong{\code{highlight.box}} when constructing a plot, but probably not ideal for the final version of a dendrogram. Defaults to FALSE}

\item{legend}{Show or hide the legend with \code{TRUE} or \code{FALSE}.}

\item{black}{Cast the color of one category (defined by its
number) as black. This setting is ideal to contrast a group
for printing in black and white.}

\item{highlight}{Highlight a category (defined by its
number) by drawing around its elements on the visualization.
In a principal components analysis, multiple circular
highlights are available to contrast sets; on a dendrogram,
only one category can be highlighted with a box.}

\item{labeling}{Defines how to label items: if setting a
character vector, define one string for each item in
\strong{df}; if setting a numeric vector (e.g, \code{1} or
\strong{2}), set it to the desired element (identified via
\code{stylo}'s use of underscores in filenames).}

\item{classing}{The class or category for each item in
\strong{df}; if not set, it defaults to the first element
before an underscore in the filename of items in the
original corpus. \code{Stylo2gg} uses classing to
distinguish items by color and shape.}

\item{shapes}{Defaults to \code{FALSE} unless
\strong{\code{labeling}} is defined. Change to toggle shapes
on the visualization instead of (or in addition to) text
labels. This is useful for printing in black and white, but
it's also helpful to distinguish among similar colors.}

\item{invert.x}{Defaults to \code{FALSE}. Change to invert
the horizontal orientation in a principal components
analysis in order to approximate some ideal visualization.
(I don't think this actually changes any understanding of
the data.)}

\item{invert.y}{Defaults to \code{FALSE}. Change to invert
the vertical orientation in a principal components analysis
in order to approximate some ideal visualization. (I don't
think this actually changes any understanding of the data.)}

\item{scaling}{Toggle the option to scale features before
running a principal components analysis. Defaults to
\code{FALSE}, except for \strong{\code{viz = "PCR"}}; for
all other principal components settings, the default is
first to normalize features by z scores, which makes
\strong{\code{scaling}} less useful.}

\item{distance.measure}{The formula used for distance in
hierarchical clustering. Defaults to \code{"delta"} to use
Burrows's formula of Manhattan distance from normalized
z-scores, but it might also be reasonable here to call
\code{"euclidean"} or some other setting, imported from
stylo.}

\item{linkage}{The linkage to be used for cluster analysis.
Defaults to \code{"ward.D"}, but \code{"complete"} might
also be a reasonable setting. Options include all those
built into R.}

\item{horiz}{Set the rotation of the dendrogram in a
hierarchical cluster analysis; defaults to \code{TRUE}}

\item{axis.labels}{Defaults to \code{FALSE} except when
\strong{\code{viz = "CA"}}. Change to \code{TRUE} to show
a distance axis for the dendrogram in a cluster analysis.}

\item{highlight.nudge}{On a highlighted dendrogram,
optionally define some extra space when a box overlaps the
edge of a label.}

\item{highlight.single}{Toggle (TRUE/FALSE) to determine whether a dendrogram's highlight should draw a single box for all of the items or individual boxes for each cluster. When using \strong{\code{highlight}}, this setting will default to \code{TRUE}; when using \strong{\code{highlight.box}}, it will default too \code{FALSE}.}

\item{show.zero}{Toggle (TRUE / FALSE) for leaving space below the lowest distance to indicate zero}

\item{highlight.box}{On a dendrogram, highlight items indicated by their item numbers (from the bottom on a horizontal dendrogram, from the left on a vertical dendrogram); it might be helpful to toggle the \strong{\code{count.labels}} parameter to
\code{TRUE} to avoid having to count large data sets.}

\item{withholding}{Specify a class or classes of texts to withhold from underlying principal components analysis before these texts are then projected into that space.}

\item{loadings.spacer}{The spacer used to replace spaces in loading words, used for multiple-word n-grams. Define it as a quoted string. Defaults to \code{"_"}.}

\item{loadings.line.color}{A string defining the lines leading to loading words. Defaults to \code{lightgray}.}

\item{loadings.word.color}{A string defining the color used to display loading words. Defaults to \code{darkgray}.}

\item{loadings.upper}{Toggle (TRUE / FALSE) to convert loadings into uppercase or to leave them alone. The default is FALSE, keeping them unconverted.}

\item{plaintext}{Toggle (TRUE / FALSE) to show text labels as \code{geom_text()} layers (the default) or as \code{geom_label()} layers (when switched to \code{FALSE}.}
}
\description{
Visualize \code{stylo} data with \code{ggplot2}.
}
\details{
Because \code{stylo2gg} builds on \code{ggplot2}, almost all
commands available to that package should work here as well,
using the plus-sign syntax documented by that package.
}
\examples{
\dontrun{
my_data <- stylo()
my_data \%>\% stylo2gg()

# Move the legend
my_data \%>\% stylo2gg() +
  theme(legend.position = "bottom")
}

}
