% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylo_log.R
\name{stylo_log}
\alias{stylo_log}
\title{Log a reproducible stylometric analysis from \code{stylo}.}
\usage{
stylo_log(stylo_object, log_label = NULL, add_dir_date = FALSE)
}
\arguments{
\item{stylo_object}{An object created from running 
\strong{\code{stylo}} on a corpus}

\item{log_label}{A label added to the directory holding
data for replication. Default value is "stylo_log".}

\item{add_dir_date}{Add the current date to the 
log_label. Default value is \code{FALSE}}
}
\description{
Log a reproducible stylometric analysis from \code{stylo}.
}
\details{
\code{stylo_log()} will add directories and files in the 
working directory to cache parameters.
}
\examples{
\dontrun{
my_data <- stylo()
my_data \%>\% stylo_log()

# Pipe directly
stylo() |> stylo_log()
}

}
