% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stylo2gg.R
\name{stylo2gg}
\alias{stylo2gg}
\title{Visualize \code{stylo} data with \code{ggplot2}.}
\usage{
stylo2gg(
  df,
  viz,
  num.features,
  title = NULL,
  caption = FALSE,
  legend,
  black = NULL,
  highlight = NULL,
  labeling,
  classing,
  shapes = FALSE,
  invert.x = FALSE,
  invert.y = FALSE,
  scaling = FALSE,
  distance.measure,
  linkage,
  horiz = TRUE,
  axis.labels = FALSE,
  highlight.nudge,
  highlight.single,
  show.zero,
  highlight.box = NULL
)
}
\arguments{
\item{df}{An object saved from running \strong{\code{stylo}}
on a corpus}

\item{viz}{A choice of visualizations, either \code{"pca"}
for principal components analysis or \code{"hc"} for
hierarchical clustering; alternatively, using
\code{"PCR"}, \code{"PCV"}, or \code{"CA"}---all
inherited from \code{stylo}---will reset a number of
defaults.}

\item{num.features}{The number of features to be used for an
analysis. By default, \code{stylo}'s settings are used, but
it is easy here limit the number to a smaller set, ordered
by frequency}

\item{title}{The title that will go on the top of a chart.
This value is inherited from \code{stylo} where possible. To
remove a title, set it to \code{NULL} or to an empty
set \strong{""}.}

\item{caption}{Defaults to \code{FALSE}, except with
certain \strong{\code{viz}} settings. Change to toggle
metadata at the bottom of a visualization.}

\item{legend}{Show or hide the legend with \code{TRUE} or \code{FALSE}.}

\item{black}{Cast the color of one category (defined by its
number) as black. This setting is ideal to contrast a group
for printing in black and white.}

\item{highlight}{Highlight a category (defined by its
number) by drawing around its elements on the visualization.
In a principal components analysis, multiple circular
highlights are available to contrast sets; on a dendrogram,
only one category can be highlighted with a box.}

\item{labeling}{Defines how to label items: if setting a
character vector, define one string for each item in
\strong{df}; if setting a numeric vector (e.g, \code{1} or
\strong{2}), set it to the desired element (identified via
\code{stylo}'s use of underscores in filenames).}

\item{classing}{The class or category for each item in
\strong{df}; if not set, it defaults to the first element
before an underscore in the filename of items in the
original corpus. \code{Stylo2gg} uses classing to
distinguish items by color and shape.}

\item{shapes}{Defaults to \code{FALSE} unless
\strong{\code{labeling}} is defined. Change to toggle shapes
on the visualization instead of (or in addition to) text
labels. This is useful for printing in black and white, but
it's also helpful to distinguish among similar colors.}

\item{invert.x}{Defaults to \code{FALSE}. Change to invert
the horizontal orientation in a principal components
analysis in order to approximate some ideal visualization.
(I don't think this actually changes any understanding of
the data.)}

\item{invert.y}{Defaults to \code{FALSE}. Change to invert
the vertical orientation in a principal components analysis
in order to approximate some ideal visualization. (I don't
think this actually changes any understanding of the data.)}

\item{scaling}{Toggle the option to scale features before
running a principal components analysis. Defaults to
\code{FALSE}, except for \strong{\code{viz = "PCR"}}; for
all other principal components settings, the default is
first to normalize features by z scores, which makes
\strong{\code{scaling}} less useful.}

\item{distance.measure}{The formula used for distance in
hierarchical clustering. Defaults to \code{"delta"} to use
Burrows's formula of Manhattan distance from normalized
z-scores, but it might also be reasonable here to call
\code{"euclidean"} or some other setting, imported from
stylo.}

\item{linkage}{The linkage to be used for cluster analysis.
Defaults to \code{"ward.D"}, but \code{"complete"} might
also be a reasonable setting. Options include all those
built into R.}

\item{horiz}{Set the rotation of the dendrogram in a
hierarchical cluster analysis; defaults to \code{TRUE}}

\item{axis.labels}{Defaults to \code{FALSE} except when
\strong{\code{viz = "CA"}}. Change to \code{TRUE} to show
a distance axis for the dendrogram in a cluster analysis.}

\item{highlight.nudge}{On a highlighted dendrogram,
optionally define some extra space when a box overlaps the
edge of a label.}

\item{highlight.single}{Toggle (TRUE/FALSE) to determine whether a dendrogram's highlight should draw a single box for all of the items or individual boxes for each cluster. When using \strong{\code{highlight}}, this setting will default to \code{TRUE}; when using \strong{\code{highlight.box}}, it will default too \code{FALSE}.}

\item{show.zero}{Toggle (TRUE / FALSE) for leaving space below the lowest distance to indicate zero}

\item{highlight.box}{On a dendrogram, highlight items indicated by their item numbers (from the bottom on a horizontal dendrogram, from the left on a vertical dendrogram)}
}
\description{
Visualize \code{stylo} data with \code{ggplot2}.
}
\details{
Because \code{stylo2gg} builds on \code{ggplot2}, almost all
commands available to that package should work here as well,
using the plus-sign syntax documented by that package.
}
\examples{
my_data <- stylo()
my_data \%>\% stylo2gg()

# Move the legend
my_data \%>\% stylo2gg() +
  theme(legend.position = "bottom")

}
